requirements.txt
aiogram==3.1.1
googletrans==3.1.0a0
python-dotenv==1.0.0